Delete Specifier.specified setter
The `specified` setter was dangerous because it made it easy to
overwrite memory unintentionally, since it provided no protection from
specifying a string that was larger than the currently specified buffer.

In its lieu, provide a `specify` shortcut that allocates a new string
and returns a specifier to it (or updates an existing specifier to point
to the new string).